pyproject.toml
[project]
name = "solar-pv-simulator"
version = "0.1.0"
description = "Streamlit-based solar PV simulation dashboard"
authors = [
    { name = "WaqasK12", email = "waaqaas.khan@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "streamlit>=1.47.1,<2.0.0",
    "streamlit-option-menu>=0.4.0,<0.5.0",
    "pillow>=11.3.0,<12.0.0",
    "openmeteo-requests",
    "retry-requests",
    "requests-cache",
    "requests",  # required by openmeteo
    "pandas",
    "numpy",
    "pytz"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
